requirements.txt
aniso8601==8.0.0
certifi==2022.9.24        
charset-normalizer==2.1.1 
click==8.0.0
colorama==0.4.5
Flask==2.1.3
Flask-RESTful==0.3.9      
Flask-SQLAlchemy==2.4.3   
Flask-WTF==1.0.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.0.0       
Jinja2==3.0.0
MarkupSafe==2.0.0rc2
numpy==1.23.3
opencv-python==4.6.0.66
Pillow==9.2.0
pybase64==1.2.3
pytz==2020.1
requests==2.28.1
six==1.15.0
SQLAlchemy==1.3.18
tk==0.1.0
torch==1.12.1
torchvision==0.13.1
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.0.0
WTForms==3.0.1
zipp==3.8.1